Initialize instance storage in cloud-init
Before, instance storage was initialized as part of the
/etc/default/docker script, which is sourced as part of
/etc/init/docker, which is run as part of starting the docker
service. This was a really wonky place to put storage
initialization!

We create a new per-boot cloud-init script containing the logic
for initializing instance storage. The old docker init
script simply does a sanity check for the presence of an expected
directory and aborts if cloud-init didn't do its job.

There is a potential regression here.

Because the old instance storage code ran as part of the docker
init routine and the docker service would retry if it failed,
we could run the instance storage code multiple times. Before,
an initially failing run could recover in a subsequent run.
Now, we will never recover because instance storage code only
runs once. This could lead to "zombie workers."

On one hand, I understand the concern. On the other, it looks like
the previous code was more paranoid than necessary with regards
to error handling. I could only see there being an issue if there
is a race condition preventing the sequence of vgcreate + lvcreate
+ mkfs.ext4 from working. I looked through Ansible's code for
lvm and filesystem management and I don't see any extra checks
after running these commands to ensure they actually worked.
I think the exit code should be all that we need to care about.
We execute the script with `set -e`, so we should be covered. One
thing we could consider doing is shutting down the instance
on failure so it doesn't become a zombie. Let's see if this
actually fails first before adding more code.