Update indexed image tasks
For indexed image pulling tests, we have a fixed task containing a test
image. Unfortunatelly, the task expired and broke the tests.

We expand the create_lz4_images script to also upload tar and zst images
and use a unique task for image pulling tests.

We also update the task IDs accordingly.