Improve robustness of retry upload tests
As we use an event catching mechanism to simulate a network down for
retry upload, during spawning of iptables process, the upload may resume
and finish before the S3 IPs are blocked.

To avoid a false negative in these tests, we retry a few times before
mark the test as failed.